+++ src/Post/PostDetail.tsx
@@ -18,6 +18,17 @@ import { Button } from '@/components/ui/button';
 import TurndownService from 'turndown';
 import MarkdownDetailSkeleton from '../Skeleton/MarkdownDetailSkeleton';
 import { ChevronLeft } from 'lucide-react';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
 
 interface PostData {
   id: number;
@@ -88,11 +99,11 @@ export default function PostDetail() {
         </ReactMarkdown>
       );
       const html = renderToStaticMarkup(markdownComponent);
-      const response = await axiosInterceptor.put(`/api/admin/s/${id}`, {
+      const response = await axiosInterceptor.put(`/api/admin/posts/${id}`, {
         title: editData.title,
         content: html,
       });
-      navigate('/documents');
+      navigate('/posts');
       toast.success('문서가 수정되었습니다.');
       console.log(response);
       await getPostDetail(markdownId!);
@@ -107,7 +118,7 @@ export default function PostDetail() {
     if (window.confirm('문서를 삭제하시겠습니까?')) {
       try {
         await axiosInterceptor.delete(`/api/admin/posts/${id}`);
-        navigate('/documents');
+        navigate('/posts');
         toast.success('문서가 삭제되었습니다.');
       } catch (error) {
         console.log(error);
@@ -210,13 +221,32 @@ export default function PostDetail() {
         <div className="flex items-center justify-between px-6">
           <CardTitle className="ck-title">체험콕 글</CardTitle>
           <div className="flex gap-3">
-            <Button
-              onClick={() => deleteMarkdown(postData.id)}
-              className="hover:bg-ck-red-500 px-4 py-2 hover:text-white"
-              variant="outline"
-            >
-              삭제
-            </Button>
+            <AlertDialog>
+              <AlertDialogTrigger asChild>
+                <Button
+                  className="hover:bg-ck-red-500 px-4 py-2 hover:text-white"
+                  variant="outline"
+                >
+                  삭제
+                </Button>
+              </AlertDialogTrigger>
+              <AlertDialogContent className="w-[350px]">
+                <AlertDialogHeader>
+                  <AlertDialogTitle>문서를 삭제하시겠습니까?</AlertDialogTitle>
+                  <AlertDialogDescription>
+                    이 작업은 되돌릴 수 없습니다
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel>취소</AlertDialogCancel>
+                  <AlertDialogAction
+                    onClick={() => deleteMarkdown(postData.id)}
+                  >
+                    확인
+                  </AlertDialogAction>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
             <Button
               onClick={() => editMarkdown(postData.id)}
               className="hover:bg-ck-blue-500 px-4 py-2 hover:text-white"
